complete task

+++ 0x06-pointers_arrays_strings/0-strcat.c
@@ -19,7 +19,7 @@ char *_strcat(char *dest, char *src)
 		i++;
 	}
 
-	for (j = 0; j < ((int) strlen(src)); j++)
+	for (j = 0; j < (strlen(src); j++)
 	{
 		dest[i] = src[j]
 	}
